% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_files.R
\name{create_par_file}
\alias{create_par_file}
\title{Generate a parameter file.}
\usage{
create_par_file(files, prefix = NULL, geno_file = NULL, snp_file = NULL,
  ind_file = NULL, badsnp_file = NULL)
}
\arguments{
\item{files}{List of filenames of the population file, parameter
file and log file.}

\item{prefix}{Prefix of the geno/snp/ind files (can include the
path). If specified, geno_file/snp_file/ind_file will be
ignored.}

\item{geno_file}{Path to the genotype file.}

\item{snp_file}{Path to the snp file.}

\item{ind_file}{Path to the ind file.}

\item{badsnp_file}{SNP file with information about ignored sites.}
}
\description{
Generate a parameter file.
}
