% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_qp3Pop_pop_file}
\alias{create_qp3Pop_pop_file}
\title{Generate a file with populations for a qp3Pop run.}
\usage{
create_qp3Pop_pop_file(A, B, C, file)
}
\arguments{
\item{A, }{B, C Population names, using the terminology of Patterson
et al., 2012. It is possible to specify vectors of population
identifiers as A, B, C. That way, all possible triplets of
populations will be generated.}

\item{file}{Path to the poplist file that will be generated.}
}
\description{
For details about the format of the population list
    file see documentation of the ADMIXTOOLS package.
}
