% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_qpAdm_pop_files}
\alias{create_qpAdm_pop_files}
\title{Generate a file with populations for a qpAdm run.}
\usage{
create_qpAdm_pop_files(L, R, files)
}
\arguments{
\item{L, }{R Sets of left (U) and right (R) populations using the
terminology of Haak et al., 2012 (Supplementary Information 10
on page 128).}

\item{files}{A list that must contain "popleft" and "popright" elements, which describe the paths to files containing "left" and "right" populations (one population per line).}
}
\description{
For details about the format of the population list
    file see documentation of the ADMIXTOOLS package.
}
