% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_qpDstat_pop_file}
\alias{create_qpDstat_pop_file}
\title{Generate a file with populations for a qpDstat run.}
\usage{
create_qpDstat_pop_file(W = NULL, X = NULL, Y = NULL, Z = NULL, file)
}
\arguments{
\item{W, }{X, Y, Z Population names, using the terminology of
Patterson et al., 2012. It is possible to specify vectors of
population identifiers as W, X, etc. That way, all possible
combinations of specified W, X, Y and Z quadruples will be
generated.}

\item{file}{Path to the poplist file that will be generated.}

\item{poplist}{List of populations. It will be saved as a file, one
population per line.}
}
\description{
For details about the format of the population list
    file see documentation of the ADMIXTOOLS package.
}
