% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixr.R
\name{create_qpF4ratio_pop_file}
\alias{create_qpF4ratio_pop_file}
\title{Generate a file with population "constellations" for a qpF4ratio run.}
\usage{
create_qpF4ratio_pop_file(X, A, B, C, O, file)
}
\arguments{
\item{X, }{A, B, C, O Population names, using the terminology of
Patterson et al., 2012}

\item{file}{Path to the poplist file that will be generated.}
}
\description{
For details about the format of the population list
    file see documentation of the ADMIXTOOLS package.
}
