% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{d}
\alias{d}
\title{Calculate D statistics or F4 statistics (which is just the
numerator of a D statistic) and return results as a data.frame.}
\usage{
d(W, X, Y, Z, prefix = NULL, geno = NULL, snp = NULL, ind = NULL,
  badsnp = NULL, outdir = NULL, f4mode = FALSE)
}
\arguments{
\item{W, X, Y, Z}{Population names according to nomenclature in Patterson
et al., 2012.}

\item{prefix}{Prefix of the geno/snp/ind files (including the whole
path).}

\item{geno, snp, ind}{Path to the geno/snp/ind file. Each overrides the 'prefix' argument.}

\item{badsnp}{SNP file with information about ignored sites.}

\item{outdir}{Where to put all generated files (temporary
directory by default).}

\item{f4mode}{Calculate f4 statistics instead of D statistic. One can also
use a dedicated f4 function.}
}
\description{
Calculate D statistics or F4 statistics (which is just the
numerator of a D statistic) and return results as a data.frame.
}
