% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adm.R
\name{doAdmixtureProjection}
\alias{doAdmixtureProjection}
\title{project genotypes on ADMIXTURE P matrix}
\usage{
doAdmixtureProjection(gt, pM)
}
\arguments{
\item{gt}{genotype vector}

\item{pM}{The P matrix of ancestral cluster frequencies from ADMIXTURE,
dimensions n SNPs x components}
}
\value{
a vector with cluster proportions
}
\description{
\code{doAdmixtureProjection} projet a genotype vector on a
given P matrix. Genotype vector is coded as 0,1,2
and has to match coding from P matrix
}

