% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fstats.R
\name{doDTest}
\alias{doDTest}
\title{D test}
\usage{
doDTest(p, idxM, blockIdx)
}
\arguments{
\item{p}{allele counts array, dimensions SNPs x individuals x 2}

\item{idxM}{matrix of dimensions 1 x 4 with population labels for test
configuration D(p1,p2)(p3,p4)}

\item{blockIdx}{index vector of chromosome blocks for jackknife}
}
\value{
a data frame with D results
\itemize{
  \item p1, p2, p3, p4  .. test configuration
  \item D .. estimate
  \item se, Z .. standard error and Z score from block jackknife
  \item nSites .. number of informative SNPs used in the test
}
}
\description{
Takes a matrix of allele frequencies as input and does D test as in
Patterson et al. 2012 on specified input populations
}

