% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_data}
\alias{download_data}
\title{Download example SNP data.}
\usage{
download_data(dirname = tempdir())
}
\arguments{
\item{dirname}{Directory in which to put the data (EIGENSTRAT trio
of snp/geno/ind files).}
}
\description{
The data is downloaded to a temporary directory by default.
}
