% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenstrat.R
\name{eigenstrat}
\alias{eigenstrat}
\title{EIGENSTRAT data constructor}
\usage{
eigenstrat(prefix)
}
\arguments{
\item{prefix}{Shared path to an EIGENSTRAT trio (set of ind/snp/geno files).}
}
\value{
S3 object of a class EIGENSTRAT
}
\description{
Construct an EIGENSTRAT S3 object from a specified path to the ind/snp/geno
trio.
}
\details{
The EIGENSTRAT data S3 object encapsulates all paths to data files required
for an analysis.
}
