% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixr.R
\name{eigenstrat_to_vcf}
\alias{eigenstrat_to_vcf}
\title{Convert EIGENSTRAT files into a VCF file.}
\usage{
eigenstrat_to_vcf(prefix, vcf_file, compress = TRUE, index = TRUE)
}
\arguments{
\item{prefix}{Prefix of the geno/snp/ind files (including the whole
path).}

\item{vcf_file}{Path to the VCF file.}

\item{compress}{Compress the VCF with bgzip?}

\item{index}{Index the VCF with tabix?}
}
\description{
This function reads the genotypes from a 'geno' file, their
coordinates and reference/alternative alleles and the individuals'
names from an 'ind' file and generates a minimalistic VCF file.
}
\details{
Compressing and indexing requires having "bgzip" and "tabix"
commands in $PATH.
}
