% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{f3}
\alias{f3}
\title{Calculate 3-population statistic.}
\usage{
f3(A, B, C, prefix = NULL, geno = NULL, snp = NULL, ind = NULL,
  badsnp = NULL, outdir = NULL, inbreed = FALSE)
}
\arguments{
\item{A, B, C}{Population names.}

\item{prefix}{Prefix of the geno/snp/ind files (including the whole
path).}

\item{geno, snp, ind}{Path to the geno/snp/ind file. Each overrides the 'prefix' argument.}

\item{badsnp}{SNP file with information about ignored sites.}

\item{outdir}{Where to put all generated files (temporary
directory by default).}

\item{inbreed}{See README.3PopTest in ADMIXTOOLS.}
}
\description{
Calculate 3-population statistic.
}
