% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{f4}
\alias{f4}
\title{Calculate F4 statistic (which is just the numerator of D statistic) and
return results as a data.frame.}
\usage{
f4(W, X, Y, Z, prefix = NULL, geno = NULL, snp = NULL, ind = NULL,
  badsnp = NULL, dir_name = NULL)
}
\arguments{
\item{W, X, Y, Z}{Population names according to nomenclature in Patterson
et al., 2012.}

\item{prefix}{Prefix of the geno/snp/ind files (including the whole
path).}

\item{geno, snp, ind}{Path to the geno/snp/ind file. Each overrides the 'prefix' argument.}

\item{badsnp}{SNP file with information about ignored sites.}

\item{dir_name}{Where to put all generated files (temporary
directory by default).}
}
\description{
Calculate F4 statistic (which is just the numerator of D statistic) and
return results as a data.frame.
}
