% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{f4ratio}
\alias{f4ratio}
\alias{d}
\alias{f4}
\alias{f3}
\title{Calculate the D, f4, f4-ratio, or f3 statistic.}
\usage{
f4ratio(X, A, B, C, O, prefix = NULL, ind_suffix = NULL,
  exclude = NULL, outdir = NULL)

d(W, X, Y, Z, prefix = NULL, ind_suffix = NULL, exclude = NULL,
  outdir = NULL, f4mode = FALSE)

f4(W, X, Y, Z, prefix = NULL, ind_suffix = NULL, exclude = NULL,
  outdir = NULL)

f3(A, B, C, prefix = NULL, ind_suffix = NULL, exclude = NULL,
  outdir = NULL, inbreed = FALSE)
}
\arguments{
\item{prefix}{An EIGENSTRAT prefix (shared by the geno/snp/ind trio).}

\item{exclude}{Path to file in a snp format with SNPs to exclude from the calculation.}

\item{outdir}{Where to put all generated files (temporary directory by default).}

\item{W, X, Y, Z, A, B, C, O}{Population names according to the nomenclature used in
Patterson et al., 2012.}

\item{f4mode}{Calculate the f4 statistic instead of the D statistic.}

\item{inbreed}{See README.3PopTest in ADMIXTOOLS for an explanation.}
}
\description{
Calculate the D, f4, f4-ratio, or f3 statistic.
}
