% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{fastTable}
\alias{fastTable}
\title{Tabulate occurences in a vector}
\usage{
fastTable(x)
}
\arguments{
\item{x}{A vector of values}
}
\value{
A named vector with occurences of the elements of x, sorted along x
}
\description{
Tabulate occurences in a vector
}
\examples{
fastTable(c(1:5, 2))
fastTable(c("D", "A", "C", "C"))
}

