% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_damage}
\alias{filter_damage}
\title{Generate a snp file with positions of transversion SNPs.}
\usage{
filter_damage(prefix, outsnp)
}
\arguments{
\item{prefix}{An EIGENSTRAT prefix.}

\item{outsnp}{Path to an output snp file with coordinates of excluded sites.}
}
\description{
The output of this function can be used as 'exclude = ' argument in all main
admixr statistics functions.
}
