% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_sites}
\alias{filter_sites}
\title{Generate coordinates of SNPs that overlap/exclude regions in BED file}
\usage{
filter_sites(prefix, bed, outsnp, remove = FALSE)
}
\arguments{
\item{prefix}{An EIGENSTRAT prefix.}

\item{bed}{Path to a BED file.}

\item{outsnp}{Path to an output snp file with coordinates of excluded sites.}

\item{remove}{Remove sites falling inside the BED file regions? By default, sites
that do not overlap BED regions are removed.}
}
\description{
This function normally produces a snp file with sites that did not overlap
regions in the specified BED file, that can be then directly supplied to
any admixr statistic function using it's 'exclude = ' argument. If the
BED file contains regions that have to be filtered out, set 'remove = TRUE'.
}
