% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{getChromosomeBlocks}
\alias{getChromosomeBlocks}
\title{Generate index vector for chromosome blocks}
\usage{
getChromosomeBlocks(chromosomes, positions, blockSize)
}
\arguments{
\item{chromosomes}{A vector of chromosome IDs}

\item{positions}{A vector of chromosomal positions}

\item{blockSize}{size of the blocks}
}
\value{
A vector with the block index for each input position
}
\description{
Takes vectors of chromosome IDs and positions as inputs, together with block size
returns an index vector corresponding to segments of desired size
}

