% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fstats.R
\name{getFstAlleleCounts}
\alias{getFstAlleleCounts}
\title{Get FST from allele counts}
\usage{
getFstAlleleCounts(p, region = FALSE, extended = FALSE)
}
\arguments{
\item{p}{allele counts array, dimensions SNPs x individuals x 2}

\item{region}{optional flag to return average FST for whole input region}

\item{extended}{optional flag to also return numerator and denominator terms for per-SNP FST}
}
\value{
\itemize{
  \item region = TRUE .. average FST over region
  \item region = FALSE, extended = FALSE .. a vector of FST per marker
  \item region = FALSE, extended = TRUE .. a matrix with numerator and denominator terms of FST, as well per marker FST
}
}
\description{
Takes an array of allele counts input and calculates FST over all populations
in the input data. This is the moment estimator of Weir and Hill (2002)
using sample allele frequncies, assumes no local inbreeding.
Options to report per SNP or whole region FST value, as well as
extended output of numerator and denominator terms
}

