% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fstats.R
\name{getFstGT}
\alias{getFstGT}
\title{Get FST from genotypes}
\usage{
getFstGT(gt, popIdx, region = FALSE, extended = FALSE)
}
\arguments{
\item{gt}{genotype matrix, dimensions SNPs x individuals}

\item{popIdx}{A vector specifying the population label for each individual (column) of the genotype matrix}

\item{region}{optional flag to return average FST for whole input region}

\item{extended}{optional flag to return all terms for per-SNP FST}
}
\value{
\itemize{
  \item region = TRUE .. average FST over region
  \item region = FALSE, extended = FALSE .. a vector of FST per marker
  \item region = FALSE, extended = TRUE .. a matrix with numerator and denominator terms of FST (a, b, c in Weir & Cockerham 1984)
}
}
\description{
Takes a matrix genotypes as input and calculates FST (Weir and Cockerham 1984)
over all populations in the input data. Options to report per SNP
or whole region FST value, as well as extended output
of individual terms for estimates
}

