% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pca.R
\name{getPcaGT}
\alias{getPcaGT}
\title{calculate PCA from a genotype matrix}
\usage{
getPcaGT(gt, idsPca = colnames(gt), pcs = 10)
}
\arguments{
\item{gt}{A matrix of genotypes, dimensions SNPs x individuals}

\item{idsPca}{A vector of IDs for individuals to use to build PCA (default: all)}

\item{pcs}{Number of PCs to calculate}
}
\value{
A list with two components
\itemize{
  \item summary .. PCs, variance explained, loadings
  \item raw .. raw results from svd
}
}
\description{
\code{getPcaGT} calculates PCA from a genotype matrix.
genotype matrix is coded as 0,1,2. includes option to infer PCA from a
subset of individuals and project the remaing ones.
loadings are calculated as in Zou et al. Hum Hered 2010 June; 70(1): 9-22
}

