% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{group_labels}
\alias{group_labels}
\title{Merge multiple samples/populations under a single label}
\usage{
group_labels(prefix, ind_suffix, labels)
}
\arguments{
\item{prefix}{EIGENSTRAT prefix.}

\item{ind_suffix}{Suffix of a modified ind file.}

\item{labels}{A named list of labels to merge.}
}
\description{
Merge multiple samples/populations under a single label
}
\examples{

\dontrun{
# This will create a new ind file with labels in the 3rd column replaced by
# Europe", "EastAfrica" and "WestAfrica", respectively.
group_labels(
    prefix = <path_to_eigenstrat>
    ind_suffix = ".populations", # will create ind file '<path_to_eigenstrat>.populations'
    labels = list(Europe = c("French", "Sardinian", "Czech"),
                  WestAfrica = c("Yoruba", "Mende"))
)
}

}
