% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixr.R
\name{merge_pops}
\alias{merge_pops}
\title{Merge populations from an EIGENSTRAT "ind" file under a single
population label.}
\usage{
merge_pops(file, modified_file, merge)
}
\arguments{
\item{file}{EIGENSTRAT ind file to modify.}

\item{modified_file}{Modified EIGENSTRAT ind filename.}

\item{merge}{List of labels to merge. List names specified labels
to merge into.}
}
\description{
Merge populations from an EIGENSTRAT "ind" file under a single
population label.
}
