% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixr.R
\name{merge_snps}
\alias{merge_snps}
\title{Merge a set of EIGENSTRAT files (one EIGENSTRAT "tripplet" per
chromosome, for example) into a single EIGENSTRAT dataset.}
\usage{
merge_snps(prefixes, output_prefix)
}
\arguments{
\item{prefixes}{A vector of prefixes of a set of EIGENSTRAT files.}

\item{output_prefix}{What prefix to use for the final merged
EIGENSTRAT dataset.}
}
\description{
This function will merge files "vertically", not "horizontally"
(i.e. merging the same set of SNPs for different samples, for
example - what Admixtools' mergeit command does). Therefore, all
specified EIGENSTRAT files must contain the same set of samples. If
they don't, this function will fail and throw an exception.
}
