% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adm.R
\name{plotAdmixture}
\alias{plotAdmixture}
\title{plot ADMIXTURE results}
\usage{
plotAdmixture(sampleId, popId, k, value, colors, kOrder = 1:max(k),
  width = 1, alpha = 1, popLabels = TRUE, labColors = "black",
  rot = 0, showLegend = FALSE)
}
\arguments{
\item{sampleId}{vector of sample IDs}

\item{popId}{vector of population IDs for grouping of samples}

\item{k}{vector of cluster indices (integer)}

\item{value}{vector of cluster proportions (integer)}

\item{colors}{vector of colors for each of the K clusters}

\item{kOrder}{vector of ordering for the K clusters}

\item{width}{vector of bar widths for cluster proportions}

\item{alpha}{vector of transparency for cluster proportions}

\item{popLabels}{use population IDs instead of sample IDs for x axis (logical)}

\item{labColors}{vector of colors to use for x axis labels}

\item{rot}{angle for x axis labels (integer)}

\item{showLegend}{display legend for cluster ID and color (logical)}
}
\value{
a ggplot2 plot object
}
\description{
\code{plotAdmixture} generates classic ADMIXTURE barplots for ancestry proportions
with various customizations. Input has to be in long format, i.e. each combination
of individual / cluster / proportion on a separate line
}

