% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fstats.R
\name{plotFPairs}
\alias{plotFPairs}
\title{Plot pairs of f stats}
\usage{
plotFPairs(x, y, seX, seY, groupId, groupColors, label = NULL, z = 3,
  size = 2, shape = 23, alpha = 0.5, labSize = 8, alphaZ = 0.5,
  showLegend = TRUE)
}
\arguments{
\item{x}{vector of estimates for x axis}

\item{y}{vector of estimates for y axis}

\item{seX}{vector of standard errors for x axis}

\item{seY}{vector of standard errors for y axis}

\item{groupId}{vector of group labels}

\item{groupColors}{named vector of colors for each group; order of groupings is taken from this}

\item{label}{vector of text labels for observations (default: no labels)}

\item{z}{Z-score cutoffs for whiskers (default: 1, 3)}

\item{size}{size for plot symbols (default: 2)}

\item{shape}{shape for plot symbols (default: 23)}

\item{alpha}{transparency for plot symbols (default: 0.5)}

\item{labSize}{size for axis and facet labels (default: 8)}

\item{alphaZ}{transparency for Z-score whiskers (default: 0.5)}

\item{showLegend}{show color legend for groups (default: TRUE)}
}
\value{
a ggplot2 plot object
}
\description{
Function to plot pairwise D/f3 stats
}

