% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fstats.R
\name{plotFStat}
\alias{plotFStat}
\title{Plot f3/D stat against population ID}
\usage{
plotFStat(value, se, label, groupId, groupColors, z = c(1, 3),
  intercept = 0, size = 2, shape = 23, labSize = 8, grouped = FALSE,
  orderValue = FALSE, horizontal = TRUE, showLegend = TRUE)
}
\arguments{
\item{value}{vector of estimates}

\item{se}{vector of standard errors}

\item{label}{vector of population labels}

\item{groupId}{vector of group labels}

\item{groupColors}{named vector of colors for each group; order of groupings is taken from this}

\item{z}{Z-score cutoffs for whiskers (default: 1, 3)}

\item{intercept}{value for lines indicating desired statistic values (default: 0)}

\item{size}{size for plot symbols (default: 2)}

\item{shape}{shape for plot symbols (default: 23)}

\item{labSize}{size for axis and facet labels (default: 8)}

\item{grouped}{combine observations in groups for plotting; order of groups given by groupColor (default: FALSE)}

\item{orderValue}{reorder observations by decreasing estimate; forces grouped = FALSE (default: FALSE)}

\item{horizontal}{plot populations on the x-axis (default: TRUE)}

\item{showLegend}{show color legend for groups (default: TRUE)}
}
\value{
a ggplot2 plot object
}
\description{
Function to plot results from D/f3 stat with possible groupings;
order of populations will be same as input order for default settings
}

