% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{print.admixr_result}
\alias{print.admixr_result}
\title{Print out the admixr result object (dataframe or a list) without showing
the hidden attributes.}
\usage{
\method{print}{admixr_result}(x, ...)
}
\arguments{
\item{x}{admixr output object (dataframe or a list produced by qpAdm/qpWave)}

\item{...}{Additional arguments passed to print.}
}
\description{
Print out the admixr result object (dataframe or a list) without showing
the hidden attributes.
}
