% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenstrat.R
\name{print.EIGENSTRAT}
\alias{print.EIGENSTRAT}
\title{EIGENSTRAT print method}
\usage{
\method{print}{EIGENSTRAT}(data)
}
\arguments{
\item{data}{EIGENSTRAT data object.}
}
\description{
Print EIGENSTRAT object components.
}
