% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{printlog}
\alias{printlog}
\title{Print full output log associated with an admixr output.}
\usage{
printlog(x, target = NA, save = FALSE, prefix = NA)
}
\arguments{
\item{x}{Output from one of the admixr wrappers}

\item{target}{Specify a specific log to examine (relevant multiple target qpAdm run)}

\item{save}{Save the log output(s) to a file?}

\item{prefix}{Path prefix to the saved log output (name of the admixr command by default)}
}
\description{
Print full output log associated with an admixr output.
}
