% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{qp3Pop}
\alias{qp3Pop}
\title{Calculate the 3-population statistic and return the results as a data.frame.}
\usage{
qp3Pop(A, B, C, prefix = NULL, geno = NULL, snp = NULL, ind = NULL,
  badsnp = NULL, dir_name = NULL, inbreed = FALSE)
}
\arguments{
\item{A}{Population names, using the terminology of Patterson (2012).}

\item{B}{Population names, using the terminology of Patterson (2012).}

\item{C}{Population names, using the terminology of Patterson (2012).}

\item{prefix}{Prefix of the geno/snp/ind files (including the whole
path).}

\item{geno}{Path to the genotype file. Overrides the 'prefix'
argument.}

\item{snp}{Path to the snp file. Overrides the 'prefix' argument.}

\item{ind}{Path to the ind file. Overrides the 'prefix' argument.}

\item{badsnp}{SNP file with information about ignored sites.}

\item{dir_name}{Where to put all generated files (temporary
directory by default).}

\item{inbreed}{See README.3PopTest in ADMIXTOOLS.}
}
\description{
Calculate the 3-population statistic and return the results as a data.frame.
}
