% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpAdm.R
\name{qpAdm_filter}
\alias{qpAdm_filter}
\title{Filter qpAdm rotation results for only 'sensible' models}
\usage{
qpAdm_filter(x, p)
}
\arguments{
\item{x}{Output of a qpAdm_rotation() function}

\item{p}{p-value cutoff}
}
\value{
qpAdm_rotation object filtered down based on p-value
}
\description{
Filter for p-value larger than a specified cuttof and admixture
proportions between 0 and 1.
}
