% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpAdm.R
\name{qpAdm_prescreen}
\alias{qpAdm_prescreen}
\title{Perform a 'pre-screening' of the 'right' populations before
a qpAdm analysis based on Harney et al., 2020 (bioRxiv)}
\usage{
qpAdm_prescreen(data, candidates, left, Zcutoff = 2)
}
\arguments{
\item{data}{EIGENSTRAT dataset}

\item{candidates}{Character vector with potential 'outgroup' populations}

\item{left}{Character vector with target and source populations}
}
\value{
Data frame of all combinations of f4(L_i, L_j; R_k, R_l),
  ordered by Zscore of the f4 statistic
}
\description{
Perform a 'pre-screening' of the 'right' populations before
a qpAdm analysis based on Harney et al., 2020 (bioRxiv)
}
