% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{qpDstat}
\alias{qpDstat}
\title{Calculate D statistics or F4 statistics (which is just the
numerator of a D statistic) and return the results as a data.frame.}
\usage{
qpDstat(W, X, Y, Z, prefix = NULL, geno = NULL, snp = NULL, ind = NULL,
  badsnp = NULL, dir_name = NULL, f4mode = FALSE)
}
\arguments{
\item{W}{Population names, using the terminology of Patterson (2012).}

\item{X}{Population names, using the terminology of Patterson (2012).}

\item{Y}{Population names, using the terminology of Patterson (2012).}

\item{Z}{Population names, using the terminology of Patterson (2012).}

\item{prefix}{Prefix of the geno/snp/ind files (including the whole
path).}

\item{geno}{Path to the genotype file. Overrides the 'prefix' argument.}

\item{snp}{Path to the snp file. Overrides the 'prefix' argument.}

\item{ind}{Path to the ind file. Overrides the 'prefix' argument.}

\item{badsnp}{SNP file with information about ignored sites.}

\item{dir_name}{Where to put all generated files (temporary
directory by default).}

\item{f4mode}{Calculate f4 statistics instead of D statistic.}
}
\description{
Calculate D statistics or F4 statistics (which is just the
numerator of a D statistic) and return the results as a data.frame.
}
