% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{qpWave}
\alias{qpWave}
\title{Find the most likely number of ancestry waves.}
\usage{
qpWave(data, left, right, maxrank = NULL, matrices = FALSE,
  outdir = NULL)
}
\arguments{
\item{data}{EIGENSTRAT data object.}

\item{left, right}{Character vectors of populations labels.}

\item{maxrank}{Maximum rank to test for.}

\item{matrices}{Return the A, B matrices used in the rank calculations?}

\item{outdir}{Where to put all generated files (temporary directory by default).}
}
\description{
Find the most likely number of ancestry waves.
}
