% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenstrat_io.R
\name{read_eigenstrat}
\alias{read_eigenstrat}
\title{Read a trio of EIGENSTRAT geno/snp/ind files.}
\usage{
read_eigenstrat(prefix = NULL)
}
\arguments{
\item{prefix}{EIGENSTRAT geno/snp/ind prefix.}
}
\value{
List of three data.frame objects.
}
\description{
This function will parse all three components of an EIGENSTRAT trio and return them
as a list.
}
\details{
Note that \code{read_geno()} will only read plain-text geno files, not compressed ones.
}
