% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_parsers.R
\name{read_output}
\alias{read_output}
\title{Read an output file from one of the ADMIXTOOLS programs.}
\usage{
read_output(file)
}
\arguments{
\item{file}{Path to a file.}
}
\value{
A data.frame with parsed results.
}
\description{
Read an output file from one of the ADMIXTOOLS programs.
}
