% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_parsers.R
\name{read_qp3Pop}
\alias{read_qp3Pop}
\title{Read output log file from a qp3Pop run.}
\usage{
read_qp3Pop(file)
}
\arguments{
\item{file}{Name of the output log file.}
}
\value{
Tibble object with parsed results.
}
\description{
Read output log file from a qp3Pop run.
}
