% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_parsers.R
\name{read_qpDstat}
\alias{read_qpDstat}
\title{Read output log file from a qpDstat run.}
\usage{
read_qpDstat(file)
}
\arguments{
\item{file}{Name of the output log file.}
}
\value{
Tibble object with parsed results.
}
\description{
Read output log file from a qpDstat run.
}
