% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixr.R
\name{read_qpF4ratio}
\alias{read_qpF4ratio}
\title{Read output log file from a qpF4ratio run.}
\usage{
read_qpF4ratio(file)
}
\arguments{
\item{file}{Name of the output log file.}
}
\value{
Tibble object with the parsed results.
}
\description{
Read output log file from a qpF4ratio run.
}
