% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenstrat_io.R
\name{read_snp}
\alias{read_snp}
\title{Read an EIGENSTRAT 'snp' file.}
\usage{
read_snp(file)
}
\arguments{
\item{file}{Path to the file.}
}
\value{
A data.frame object with SNP data (columns defined by the
    EIGENSTRAT format).
}
\description{
Read an EIGENSTRAT 'snp' file.
}
