% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenstrat.R
\name{relabel}
\alias{relabel}
\title{Change labels of populations or samples}
\usage{
relabel(data, labels, outfile = tempfile(fileext = ".ind"))
}
\arguments{
\item{data}{EIGENSTRAT trio.}

\item{labels}{A named list of labels to merge (each new group defined as a
character vector).}

\item{outfile}{Path to an output snp file with coordinates of excluded sites.}
}
\value{
Updated S3 EIGENSTRAT data object with an additional 'group' slot
    specifying the path to a new ind file that will be used in downstream
    analysis.
}
\description{
Replace population/sample names with specified group labels.
}
