% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenstrat.R
\name{remove_transitions}
\alias{remove_transitions}
\title{Filter out transitions (C->T and G->A substitutions)}
\usage{
remove_transitions(data, outfile = tempfile(fileext = ".snp"))
}
\arguments{
\item{data}{EIGENSTRAT data object.}

\item{outfile}{Path to an output snp file with coordinates of excluded sites.}
}
\value{
Updated S3 EIGENSTRAT data object with an additional 'exclude' slot
    specifying the path to the set of SNPs to be removed from a downstream
    analysis.
}
\description{
Remove substitutions that are more likely to be a result of ancient DNA
damage (C->T and G->A substitutions).
}
