% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenstrat.R
\name{reset}
\alias{reset}
\title{Reset modifications to an EIGENSTRAT object}
\usage{
reset(data)
}
\arguments{
\item{data}{EIGENSTRAT data object.}
}
\value{
EIGENSTRAT data S3 object.
}
\description{
Set 'exclude' and 'group' modifications of snp and ind files, respectively,
to NULL.
}
