% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_cmd}
\alias{run_cmd}
\title{Run a given ADMIXTOOLS command.}
\usage{
run_cmd(cmd, par_file, log_file)
}
\arguments{
\item{par_file}{Path to the parameter file.}

\item{log_file}{Path to the output file. If NULL, output will be
printed to stdout.}

\item{tool}{Which ADMIXTOOLS command to run. At the moment, only
qpDstat and qpF4ratio are supported.}
}
\description{
Run a given ADMIXTOOLS command.
}
