% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{snps_missing}
\alias{snps_missing}
\title{Calculate the number (or proportion) of sites with an allele
missing for each sample.}
\usage{
snps_missing(geno, prop = FALSE)
}
\arguments{
\item{geno}{EIGENSTRAT geno dataframe.}

\item{prop}{Calculate the proportion of missing alleles
instead.}
}
\value{
A named vector of counts or proportions.
}
\description{
Calculate the number (or proportion) of sites with an allele
missing for each sample.
}
