% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixr.R
\name{snps_present}
\alias{snps_present}
\title{Calculate the number (or proportion) of sites with an allele
present (i.e. not 9) for each sample.}
\usage{
snps_present(geno, prop = FALSE)
}
\arguments{
\item{geno}{EIGENSTRAT geno dataframe.}

\item{prop}{Calculate the proportion of non-missing alleles
instead.}
}
\value{
A named vector of counts or proportions.
}
\description{
Calculate the number (or proportion) of sites with an allele
present (i.e. not 9) for each sample.
}
