% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_sites}
\alias{subset_sites}
\title{Subset data in an EIGENSTRAT file based on a given BED file.}
\usage{
subset_sites(prefix, subset_prefix, bed_file, complement = FALSE)
}
\arguments{
\item{prefix}{Prefix of EIGENSTRAT geno/snp/ind files.}

\item{subset_prefix}{Prefix of the EIGENSTRAT subset.}

\item{bed_file}{Path to a 3 column BED file to intersect with.}

\item{complement}{Perform an intersect or a complement operation?}
}
\description{
Subset data in an EIGENSTRAT file based on a given BED file.
}
