% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixr.R
\name{write_geno}
\alias{write_geno}
\title{Write an EIGENSTRAT 'geno' file.}
\usage{
write_geno(geno_file, df)
}
\arguments{
\item{file}{Path to the file.}

\item{geno}{data.frame with data in a 'geno' format}
}
\description{
Write an EIGENSTRAT 'geno' file.
}
