% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixr.R
\name{write_ind}
\alias{write_ind}
\title{Write an EIGENSTRAT 'ind' file.}
\usage{
write_ind(ind_file, df)
}
\arguments{
\item{file}{Path to the file.}

\item{ind}{data.frame with data in an 'ind' format}
}
\description{
Write an EIGENSTRAT 'ind' file.
}
