% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_snps}
\alias{count_snps}
\title{Calculate the number/proportion of covered sites in each sample.}
\usage{
count_snps(prefix, missing = FALSE, prop = FALSE)
}
\arguments{
\item{prefix}{EIGENSTRAT prefix.}

\item{missing}{Count present SNPs or missing SNPs?}

\item{prop}{Calculate the proportion of non-missing alleles instead?}
}
\value{
data.frame object with SNP counts/proportions.
}
\description{
Calculate the number/proportion of covered sites in each sample.
}
