% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_par_file}
\alias{create_par_file}
\title{Generate a parameter file.}
\usage{
create_par_file(par_file, pop_file, eigenstrat_prefix = NULL,
  geno_file = NULL, snp_file = NULL, ind_file = NULL,
  badsnp_file = NULL, f4mode = FALSE)
}
\arguments{
\item{par_file}{Path to the parameter file generated by this
function.}

\item{pop_file}{Populations file (qpDstat quadruples, Dstats
population list, qpF4ratio populations, etc.).}

\item{eigenstrat_prefix}{Prefix of the geno/snp/ind files (can
include the path). If specified, geno_file/snp_file/ind_file
will be ignored.}

\item{geno_file}{Path to the genotype file.}

\item{snp_file}{Path to the snp file.}

\item{ind_file}{Path to the ind file.}

\item{badsnp_file}{SNP file with information about ignored sites.}

\item{f4mode}{Run Dstats with an "f4mode: YES" option, estimating
just the f4?}
}
\description{
Generate a parameter file.
}
