% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{f4ratio}
\alias{f4ratio}
\title{Perform an f4-ratio calculation and return its result as a data.frame.}
\usage{
f4ratio(X, A, B, C, O, prefix = NULL, geno = NULL, snp = NULL,
  ind = NULL, badsnp = NULL, outdir = NULL)
}
\arguments{
\item{X, A, B, C, O}{Population names according to nomenclature used in
Patterson et al., 2012.}

\item{prefix}{Prefix of the geno/snp/ind files (including the whole
path).}

\item{geno, snp, ind}{Path to the geno/snp/ind file. Each overrides the 'prefix' argument.}

\item{badsnp}{SNP file with information about ignored sites.}

\item{outdir}{Where to put all generated files (temporary
directory by default).}
}
\description{
Perform an f4-ratio calculation and return its result as a data.frame.
}
