% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{f4ratio}
\alias{f4ratio}
\alias{d}
\alias{f4}
\alias{f3}
\title{Calculate the D, f4, f4-ratio, or f3 statistic.}
\usage{
f4ratio(data, X, A, B, C, O, outdir = NULL)

d(data, W, X, Y, Z, outdir = NULL, f4mode = FALSE)

f4(data, W, X, Y, Z, outdir = NULL)

f3(data, A, B, C, outdir = NULL, inbreed = FALSE)
}
\arguments{
\item{data}{EIGENSTRAT data object.}

\item{outdir}{Where to put all generated files (temporary directory by default).}

\item{W, X, Y, Z, A, B, C, O}{Population names according to the nomenclature used in
Patterson et al., 2012.}

\item{f4mode}{Calculate the f4 statistic instead of the D statistic.}

\item{inbreed}{See README.3PopTest in ADMIXTOOLS for an explanation.}
}
\description{
Calculate the D, f4, f4-ratio, or f3 statistic.
}
