% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{group_labels}
\alias{group_labels}
\title{Merge multiple samples/populations under a single label}
\usage{
group_labels(ind, modified_ind, labels)
}
\arguments{
\item{ind}{Path to an original ind file.}

\item{modified_ind}{Path to a final modified ind file.}

\item{labels}{A named list of labels to merge.}
}
\description{
Merge multiple samples/populations under a single label
}
\examples{

\dontrun{
# This will create a new ind file with labels in the 3rd column replaced by
# Europe", "EastAfrica" and "WestAfrica", respectively.
group_labels(
    ind = IND_FILE,
    modified_ind = paste0(IND_FILE, ".merged"),
    labels = list(Europe = c("French", "Sardinian", "Czech"),
                  WestAfrica = c("Yoruba", "Mende"))
)
}

}
