% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_eigenstrat}
\alias{merge_eigenstrat}
\title{Merge two sets of EIGENSTRAT datasets (utilizing the 'mergeit' command).}
\usage{
merge_eigenstrat(prefix, input1, input2)
}
\arguments{
\item{prefix}{Prefix of the merged dataset.}

\item{input1, input2}{Prefixes of two EIGENSTRAT datasets to merge.}
}
\description{
Merge two sets of EIGENSTRAT datasets (utilizing the 'mergeit' command).
}
