% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenstrat.R
\name{merge_eigenstrat}
\alias{merge_eigenstrat}
\title{Merge two sets of EIGENSTRAT datasets}
\usage{
merge_eigenstrat(merged, a, b, strandcheck = "NO")
}
\arguments{
\item{merged}{Prefix of the path to the merged EIGENSTRAT snp/ind/geno trio.}

\item{a, b}{Two EIGENSTRAT objects to merge.}

\item{strandcheck}{Deal with potential strand issues? Mostly for historic reasons. For details see the README of ADMIXTOOLS convertf.}
}
\description{
This function utilizes the 'mergeit' command distributed in ADMIXTOOLS.
}
