% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_labels}
\alias{merge_labels}
\title{Merge populations from an EIGENSTRAT "ind" file under a single
population label.}
\usage{
merge_labels(ind, modified_ind, pop_labels)
}
\arguments{
\item{ind}{EIGENSTRAT ind file to modify.}

\item{modified_ind}{Modified EIGENSTRAT ind filename.}

\item{pop_labels}{Named list of labels to merge. Names specify labels
to merge into.}
}
\description{
Merge populations from an EIGENSTRAT "ind" file under a single
population label.
}
