% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_labels}
\alias{merge_labels}
\title{Merge populations from an EIGENSTRAT "ind" file under a single
population label.}
\usage{
merge_labels(ind, modified_ind, pop_labels)
}
\arguments{
\item{ind}{EIGENSTRAT ind file to modify.}

\item{modified_ind}{Modified EIGENSTRAT ind filename.}

\item{pop_labels}{Named list of labels to merge. Names specify labels
to merge into.}
}
\description{
Merge populations from an EIGENSTRAT "ind" file under a single
population label.
}
\examples{

\dontrun{
# This will create a new ind file with labels in the 3rd column replaced by
# Europe", "EastAfrica" and "WestAfrica", respectively.
merge_labels(
    ind = IND_FILE,
    modified_ind = paste0(IND_FILE, ".merged"),
    pop_labels = list(Europe = c("French", "Sardinian", "Czech"),
                      EastAfrica = "Dinka",
                      WestAfrica = c("Mbuti", "Mende"))
)
}

}
