% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_labels}
\alias{merge_labels}
\title{Merge multiple samples/populations under a single label.}
\usage{
merge_labels(ind, modified_ind, labels)
}
\arguments{
\item{ind}{EIGENSTRAT ind file to modify.}

\item{modified_ind}{Modified EIGENSTRAT ind filename.}

\item{labels}{Named list of labels to merge. Names specify labels
to merge into.}
}
\description{
Merge multiple samples/populations under a single label.
}
\examples{

\dontrun{
# This will create a new ind file with labels in the 3rd column replaced by
# Europe", "EastAfrica" and "WestAfrica", respectively.
merge_labels(
    ind = IND_FILE,
    modified_ind = paste0(IND_FILE, ".merged"),
    labels = list(Europe = c("French", "Sardinian", "Czech"),
                      WestAfrica = c("Yoruba", "Mende"))
)
}

}
