% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenstrat.R
\name{print.EIGENSTRAT}
\alias{print.EIGENSTRAT}
\title{EIGENSTRAT print method}
\usage{
\method{print}{EIGENSTRAT}(x, ...)
}
\arguments{
\item{x}{EIGENSTRAT data object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print EIGENSTRAT object components.
}
