% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{printlog}
\alias{printlog}
\title{Print full output log associated with an admixr output.}
\usage{
printlog(x)
}
\arguments{
\item{x}{Output from one of the admixr wrappers (data.frame)}
}
\description{
Print full output log associated with an admixr output.
}
