% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{qpAdm}
\alias{qpAdm}
\title{Run the qpAdm analysis.}
\usage{
qpAdm(L, R, prefix = NULL, geno = NULL, snp = NULL, ind = NULL,
  badsnp = NULL, dir_name = NULL)
}
\arguments{
\item{L}{Sets of "left" populations using the terminology of Haak (2012,
Supplementary Information 10 on page 128).}

\item{R}{Sets of "right" populations using the terminology of Haak (2012,
Supplementary Information 10 on page 128).}

\item{prefix}{Prefix of the geno/snp/ind files (including the whole
path).}

\item{geno}{Path to the genotype file. Overrides the 'prefix'
argument.}

\item{snp}{Path to the snp file. Overrides the 'prefix' argument.}

\item{ind}{Path to the ind file. Overrides the 'prefix' argument.}

\item{badsnp}{SNP file with information about ignored sites.}

\item{dir_name}{Where to put all generated files (temporary
directory by default).}
}
\description{
Run the qpAdm analysis.
}
