% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{qpAdm}
\alias{qpAdm}
\title{Calculate ancestry proportions in a set of target populations.}
\usage{
qpAdm(data, target, references, outgroups, details = FALSE,
  outdir = NULL)
}
\arguments{
\item{data}{EIGENSTRAT data object.}

\item{target}{Vector of target populations (evaluated one at a time).}

\item{references}{Reference source populations related to true ancestors.}

\item{outgroups}{Outgroup populations.}

\item{outdir}{Where to put all generated files (temporary directory by default).}
}
\description{
Calculate ancestry proportions in a set of target populations.
}
