% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{qpAdm}
\alias{qpAdm}
\title{Calculate admixture proportions in a target population using qpAdm method.}
\usage{
qpAdm(target, source, outgroup, prefix = NULL, geno = NULL,
  snp = NULL, ind = NULL, badsnp = NULL, dir_name = NULL)
}
\arguments{
\item{target}{Vector of target populations (evaluated one at a time).}

\item{source}{Source populations related to true ancestors.}

\item{outgroup}{Outgroup populations.}

\item{prefix}{Prefix of the geno/snp/ind files (including the whole
path).}

\item{geno, snp, ind}{Path to the geno/snp/ind file. Each overrides the 'prefix' argument.}

\item{badsnp}{SNP file with information about ignored sites.}

\item{dir_name}{Where to put all generated files (temporary
directory by default).}
}
\description{
Calculate admixture proportions in a target population using qpAdm method.
}
