% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{qpAdm}
\alias{qpAdm}
\title{Calculate admixture proportions in target populations using qpAdm.}
\usage{
qpAdm(target, refs, outgroups, prefix = NULL, geno = NULL,
  snp = NULL, ind = NULL, badsnp = NULL, outdir = NULL)
}
\arguments{
\item{target}{Vector of target populations (evaluated one at a time).}

\item{refs}{Reference source populations related to true ancestors.}

\item{outgroups}{Outgroup populations.}

\item{prefix}{Prefix of the geno/snp/ind files (including the whole
path).}

\item{geno, snp, ind}{Path to the geno/snp/ind file. Each overrides the 'prefix' argument.}

\item{badsnp}{SNP file with information about ignored sites.}

\item{outdir}{Where to put all generated files (temporary
directory by default).}
}
\description{
Calculate admixture proportions in target populations using qpAdm.
}
