% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{qpAdm}
\alias{qpAdm}
\title{Calculate ancestry proportions in target populations using qpAdm.}
\usage{
qpAdm(target, references, outgroups, prefix = NULL, geno = NULL,
  snp = NULL, ind = NULL, exclude = NULL, outdir = NULL)
}
\arguments{
\item{target}{Vector of target populations (evaluated one at a time).}

\item{references}{Reference source populations related to true ancestors.}

\item{outgroups}{Outgroup populations.}

\item{prefix}{An EIGENSTRAT prefix (shared by the geno/snp/ind trio).}

\item{geno, snp, ind}{Paths to individual EIGENSTRAT files. Each overrides the 'prefix' argument.}

\item{exclude}{Path to file in a snp format with SNPs to exclude from the calculation.}

\item{outdir}{Where to put all generated files (temporary directory by default).}
}
\description{
Calculate ancestry proportions in target populations using qpAdm.
}
