% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{qpF4ratio}
\alias{qpF4ratio}
\title{Perform an f4-ratio calculation and return its result as a data.frame.}
\usage{
qpF4ratio(X, A, B, C, O, prefix = NULL, geno = NULL, snp = NULL,
  ind = NULL, badsnp = NULL, dir_name = NULL)
}
\arguments{
\item{X}{Population names.}

\item{A}{Population names.}

\item{B}{Population names.}

\item{C}{Population names.}

\item{O}{Population names.}

\item{prefix}{Prefix of the geno/snp/ind files (including the whole
path).}

\item{geno}{Path to the genotype file. Overrides the 'prefix' argument.}

\item{snp}{Path to the snp file. Overrides the 'prefix' argument.}

\item{ind}{Path to the ind file. Overrides the 'prefix' argument.}

\item{badsnp}{SNP file with information about ignored sites.}

\item{dir_name}{Where to put all generated files (temporary
directory by default).}
}
\description{
Perform an f4-ratio calculation and return its result as a data.frame.
}
