% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenstrat_io.R
\name{read_eigenstrat}
\alias{read_eigenstrat}
\title{Read a trio of EIGENSTRAT geno/snp/ind files.}
\usage{
read_eigenstrat(prefix = NULL)
}
\arguments{
\item{prefix}{EIGENSTRAT geno/snp/ind prefix.}
}
\value{
List of three data.frame objects.
}
\description{
Read a trio of EIGENSTRAT geno/snp/ind files.
}
