% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenstrat_handling.R
\name{read_eigenstrat}
\alias{read_eigenstrat}
\title{Read a tripplet of EIGENSTRAT (geno/snp/ind files) files.}
\usage{
read_eigenstrat(prefix = NULL)
}
\arguments{
\item{prefix}{EIGENSTRAT prefix of geno/snp/ind files.}
}
\value{
List of three data frames (one element for geno/snp/ind).
}
\description{
Read a tripplet of EIGENSTRAT (geno/snp/ind files) files.
}
