% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixr.R
\name{read_eigenstrat}
\alias{read_eigenstrat}
\title{Read a tripplet of EIGENSTRAT (geno/snp/ind files) files.}
\usage{
read_eigenstrat(prefix = NULL, snp_file = NULL, geno_file = NULL,
  ind_file = NULL)
}
\arguments{
\item{file}{Path to the file.}
}
\value{
List of three data frames (one element for geno/snp/ind).
}
\description{
Read a tripplet of EIGENSTRAT (geno/snp/ind files) files.
}
