% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenstrat_io.R
\name{read_geno}
\alias{read_geno}
\title{Read an EIGENSTRAT 'geno' file.}
\usage{
read_geno(file, ind_file = stringr::str_replace(file, "geno$", "ind"))
}
\arguments{
\item{file}{Path to the geno file.}

\item{ind_file}{Path to the ind file to read sample names from.}
}
\value{
A data.frame object with genotypes of each sample (0/1/9 as defined
    by the EIGENSTRAT format).
}
\description{
Read an EIGENSTRAT 'geno' file.
}
