% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixr.R
\name{read_geno}
\alias{read_geno}
\title{Read an EIGENSTRAT 'geno' file.}
\usage{
read_geno(file, inds = NULL)
}
\arguments{
\item{file}{Path to the file.}
}
\value{
Data frame with columns containing "genotypes" of each
    sample (0/1/9 as defined by the EIGENSTRAT format).
}
\description{
Read an EIGENSTRAT 'geno' file.
}
