% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenstrat_io.R
\name{read_ind}
\alias{read_ind}
\title{Read an EIGENSTRAT 'ind' file.}
\usage{
read_ind(file)
}
\arguments{
\item{file}{Path to the file.}
}
\value{
A data.frame object with sample identifier, sex and label columns
    (as defined by the EIGENSTRAT format).
}
\description{
Read an EIGENSTRAT 'ind' file.
}
