% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenstrat_io.R
\name{read_ind}
\alias{read_ind}
\alias{read_snp}
\alias{read_geno}
\title{Read an EIGENSTRAT ind/snp/geno file.}
\usage{
read_ind(file)

read_snp(file)

read_geno(file)
}
\arguments{
\item{file}{Path to a file.}
}
\value{
A data.frame object.
}
\description{
These functions each read one part of the EIGENSTRAT dataset trio.
}
\details{
Note that \code{read_geno()} will only read plain-text geno files, not compressed ones.
}
