% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixr.R
\name{read_ind}
\alias{read_ind}
\title{Read an EIGENSTRAT 'ind' file.}
\usage{
read_ind(file)
}
\arguments{
\item{file}{Path to the file.}
}
\value{
Data frame with the sample identifier, sex and label
    columns.
}
\description{
Read an EIGENSTRAT 'ind' file.
}
