% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_ind}
\alias{read_ind}
\alias{read_snp}
\alias{read_geno}
\title{Read an EIGENSTRAT ind/snp/geno file.}
\usage{
read_ind(data)

read_snp(data, exclude = FALSE)

read_geno(data)
}
\arguments{
\item{data}{EIGENSTRAT data object.}

\item{exclude}{Read the list of excluded SNPs?}
}
\value{
A data.frame object.
}
\description{
These functions each read one part of the EIGENSTRAT dataset trio.
}
\details{
Note that \code{read_geno()} will only read plain-text geno files, not compressed ones.
}
