% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_parsers.R
\name{read_output}
\alias{read_output}
\title{Read output file of one of the ADMIXTOOLS' programs.}
\usage{
read_output(file)
}
\arguments{
\item{file}{Name of the output log file.}
}
\value{
data.frame with parsed results.
}
\description{
This function servers as a dispatcher delegating the parsing to
read_qpF4ratio, read_qpDstat, etc.
}
