% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_parsers.R
\name{read_output}
\alias{read_output}
\title{Read an output file from one of the ADMIXTOOLS programs.}
\usage{
read_output(file, ...)
}
\arguments{
\item{file}{A path to an output file.}

\item{...}{See the 'details' argument of qpWave.}
}
\value{
A tibble with the results.
}
\description{
Read an output file from one of the ADMIXTOOLS programs.
}
