% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixr.R, R/io.R
\name{read_snp}
\alias{read_snp}
\alias{read_snp}
\title{Read an EIGENSTRAT 'snp' file.}
\usage{
read_snp(snp_file)

read_snp(snp_file)
}
\arguments{
\item{file}{Path to the file.}

\item{file}{Path to the file.}
}
\value{
Data frame with information about each SNP (columns defined
    by the EIGENSTRAT format).

Data frame with information about each SNP (columns defined
    by the EIGENSTRAT format).
}
\description{
Read an EIGENSTRAT 'snp' file.

Read an EIGENSTRAT 'snp' file.
}
