% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{snps_present}
\alias{snps_present}
\title{Calculate the number/proportion of covered sites in each sample.}
\usage{
snps_present(prefix, prop = FALSE)
}
\arguments{
\item{prefix}{EIGENSTRAT prefix.}

\item{prop}{Calculate the proportion of non-missing alleles instead?}
}
\value{
data.frame object with SNP counts/proportions.
}
\description{
Calculate the number/proportion of covered sites in each sample.
}
