% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixr.R
\name{subset_sites}
\alias{subset_sites}
\title{Create a new set of EIGENSTRAT files by intersecting the original
data with a given set of coordinates.}
\usage{
subset_sites(prefix, out_prefix, bed_file = NULL, pos_file = NULL,
  complement = FALSE)
}
\arguments{
\item{prefix}{Prefix of the geno/snp/ind files (including the whole
path).}

\item{out_prefix}{Prefix of the generated EIGENSTRAT files with the
subset of the data.}

\item{bed_file}{Path to the 3 column BED file to intersect with.}

\item{pos_file}{Path to the 2 column position file to intersect with.}

\item{complement}{Perform an intersect or a complement operation?}
}
\description{
Create a new set of EIGENSTRAT files by intersecting the original
data with a given set of coordinates.
}
