% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vcf_to_eigenstrat}
\alias{vcf_to_eigenstrat}
\title{Convert VCF file into an EIGENSTRAT format}
\usage{
vcf_to_eigenstrat(vcf, eigenstrat)
}
\arguments{
\item{vcf}{A path to a VCF file.}

\item{eigenstrat}{A path to an output EIGENSTRAT triplet.}
}
\description{
Convert VCF file into an EIGENSTRAT format
}
