% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_io.R
\name{vcf_to_eigenstrat}
\alias{vcf_to_eigenstrat}
\title{Convert VCF file into three-file EIGENSTRAT format.}
\usage{
vcf_to_eigenstrat(vcf_file, prefix)
}
\arguments{
\item{vcf_file}{Path to the VCF file.}

\item{prefix}{Prefix of the geno/snp/ind files (including the whole
path) that will be generated.}
}
\description{
Convert VCF file into three-file EIGENSTRAT format.
}
