% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_eigenstrat.R
\name{vcf_to_eigenstrat}
\alias{vcf_to_eigenstrat}
\alias{eigenstrat_to_vcf}
\title{Conversion between VCF and EIGENSTRAT file formats}
\usage{
vcf_to_eigenstrat(vcf, prefix)

eigenstrat_to_vcf(prefix, vcf, compress = TRUE, index = TRUE)
}
\arguments{
\item{vcf}{Path to a VCF file.}

\item{prefix}{EIGENSTRAT geno/snp/ind output prefix.}

\item{compress}{Compress the VCF with bgzip?}

\item{index}{Index the VCF with tabix?}
}
\description{
These functions parse input in either VCF or EIGENSTRAT format and convert it.
}
\details{
Please note that processing of large data sets (such as whole-genome data)
using these functions will be very inefficient, because they first load all
data into memory first, perform the conversion, and only then save the
output file.

Compressing and indexing is performed by default, and requires having
bgzip and tabix commands in $PATH.
}
