% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenstrat_io.R
\name{write_eigenstrat}
\alias{write_eigenstrat}
\title{Write a tripplet of EIGENSTRAT (geno/snp/ind files) files.}
\usage{
write_eigenstrat(prefix, ind, snp, geno)
}
\arguments{
\item{prefix}{Prefix of the geno/snp/ind files (can
include the path).}

\item{ind}{data.frame with data in a 'ind' format}

\item{snp}{data.frame with data in a 'snp' format}

\item{geno}{data.frame with data in a 'geno' format}
}
\value{
List of three data frames (one element for geno/snp/ind).
}
\description{
Write a tripplet of EIGENSTRAT (geno/snp/ind files) files.
}
