% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenstrat_io.R
\name{write_eigenstrat}
\alias{write_eigenstrat}
\title{Write a trio of EIGENSTRAT geno/snp/ind files.}
\usage{
write_eigenstrat(prefix, ind, snp, geno)
}
\arguments{
\item{prefix}{EIGENSTRAT geno/snp/ind prefix.}

\item{ind, snp, geno}{EIGENSTRAT data as data.frames.}
}
\description{
Write a trio of EIGENSTRAT geno/snp/ind files.
}
