% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenstrat_io.R
\name{write_geno}
\alias{write_geno}
\title{Write an EIGENSTRAT 'geno' file.}
\usage{
write_geno(df, file)
}
\arguments{
\item{df}{Data frame with columns containing "genotypes" of each
sample (0/1/9 as defined by the EIGENSTRAT format).}

\item{file}{Path to the file.}
}
\description{
Write an EIGENSTRAT 'geno' file.
}
