% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixr.R, R/io.R
\name{write_ind}
\alias{write_ind}
\alias{write_ind}
\title{Write an EIGENSTRAT 'ind' file.}
\usage{
write_ind(ind_file, df)

write_ind(ind_file, df)
}
\arguments{
\item{file}{Path to the file.}

\item{ind}{Data frame with the sample identifier, sex and label
columns (columns defined by the EIGENSTRAT format).}

\item{file}{Path to the file.}

\item{ind}{Data frame with the sample identifier, sex and label
columns (columns defined by the EIGENSTRAT format).}
}
\description{
Write an EIGENSTRAT 'ind' file.

Write an EIGENSTRAT 'ind' file.
}
