% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenstrat_io.R
\name{write_ind}
\alias{write_ind}
\title{Write an EIGENSTRAT 'ind' file.}
\usage{
write_ind(df, file)
}
\arguments{
\item{df}{A data.frame object with genotypes of each sample (0/1/9 as defined
by the EIGENSTRAT format).}

\item{file}{Path to the file.}
}
\description{
Write an EIGENSTRAT 'ind' file.
}
