% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_ind}
\alias{write_ind}
\alias{write_snp}
\alias{write_geno}
\title{Write an EIGENSTRAT ind/snp/geno file.}
\usage{
write_ind(df, file)

write_snp(df, file)

write_geno(df, file)
}
\arguments{
\item{df}{A data.frame object.}

\item{file}{Path to an output file.}
}
\description{
Write an EIGENSTRAT ind/snp/geno file.
}
