% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenstrat_io.R
\name{write_snp}
\alias{write_snp}
\title{Write an EIGENSTRAT 'snp' file.}
\usage{
write_snp(df, file)
}
\arguments{
\item{df}{A data.frame object with SNP data (columns defined by the
EIGENSTRAT format).}

\item{file}{Path to the file.}
}
\description{
Write an EIGENSTRAT 'snp' file.
}
