% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenstrat_handling.R
\name{write_snp}
\alias{write_snp}
\title{Write an EIGENSTRAT 'snp' file.}
\usage{
write_snp(file, df)
}
\arguments{
\item{file}{Path to the file.}

\item{df}{Data frame with information about each SNP (columns
defined by the EIGENSTRAT format).}
}
\description{
Write an EIGENSTRAT 'snp' file.
}
