% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixr.R
\name{write_snp}
\alias{write_snp}
\title{Write an EIGENSTRAT 'snp' file.}
\usage{
write_snp(snp_file, df)
}
\arguments{
\item{file}{Path to the file.}

\item{snp}{Data frame with information about each SNP (columns
defined by the EIGENSTRAT format).}
}
\description{
Write an EIGENSTRAT 'snp' file.
}
